moved local api server creation to NewClientOrDie()
