+++ cmd/argocd/commands/logout.go
@@ -2,12 +2,12 @@ package commands
 
 import (
 	"fmt"
+	argocdclient "github.com/argoproj/argo-cd/v2/pkg/apiclient"
 	"os"
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
-	argocdclient "github.com/argoproj/argo-cd/v2/pkg/apiclient"
 	"github.com/argoproj/argo-cd/v2/util/errors"
 	"github.com/argoproj/argo-cd/v2/util/localconfig"
 )
