+++ cmd/argocd/commands/version.go
@@ -3,13 +3,14 @@ package commands
 import (
 	"context"
 	"fmt"
-
+	"github.com/argoproj/argo-cd/v2/cmd/argocd/commands/initialize"
+	argocdclient "github.com/argoproj/argo-cd/v2/pkg/apiclient"
+	"github.com/argoproj/argo-cd/v2/pkg/apiclient/headless"
 	"github.com/golang/protobuf/ptypes/empty"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
 	"github.com/argoproj/argo-cd/v2/common"
-	argocdclient "github.com/argoproj/argo-cd/v2/pkg/apiclient"
 	"github.com/argoproj/argo-cd/v2/pkg/apiclient/version"
 	"github.com/argoproj/argo-cd/v2/util/errors"
 	argoio "github.com/argoproj/argo-cd/v2/util/io"
@@ -40,7 +41,7 @@ func NewVersionCmd(clientOpts *argocdclient.ClientOptions) *cobra.Command {
 `,
 		Run: func(cmd *cobra.Command, args []string) {
 			cv := common.GetVersion()
-
+			ctxStr := initialize.RetrieveContextIfChanged(cmd.Flag("context"))
 			switch output {
 			case "yaml", "json":
 				v := make(map[string]interface{})
@@ -52,7 +53,7 @@ func NewVersionCmd(clientOpts *argocdclient.ClientOptions) *cobra.Command {
 				}
 
 				if !client {
-					sv := getServerVersion(clientOpts)
+					sv := getServerVersion(clientOpts, ctxStr)
 
 					if short {
 						v["server"] = map[string]string{"argocd-server": sv.Version}
@@ -67,7 +68,7 @@ func NewVersionCmd(clientOpts *argocdclient.ClientOptions) *cobra.Command {
 				printClientVersion(&cv, short || (output == "short"))
 
 				if !client {
-					sv := getServerVersion(clientOpts)
+					sv := getServerVersion(clientOpts, ctxStr)
 					printServerVersion(sv, short || (output == "short"))
 				}
 			default:
@@ -81,8 +82,8 @@ func NewVersionCmd(clientOpts *argocdclient.ClientOptions) *cobra.Command {
 	return &versionCmd
 }
 
-func getServerVersion(options *argocdclient.ClientOptions) *version.VersionMessage {
-	conn, versionIf := argocdclient.NewClientOrDie(options).NewVersionClientOrDie()
+func getServerVersion(options *argocdclient.ClientOptions, ctxStr string) *version.VersionMessage {
+	conn, versionIf := headless.NewClientOrDie(options, ctxStr).NewVersionClientOrDie()
 	defer argoio.Close(conn)
 
 	v, err := versionIf.Version(context.Background(), &empty.Empty{})
